Cleanup lsp extensions on the client side
